Move DataType and Value
