+++ automerge-backend/src/patch.rs
@@ -1,8 +1,7 @@
 use crate::op_handle::OpHandle;
-use crate::protocol::Value;
 use serde::ser::SerializeMap;
 use serde::{Deserialize, Serialize, Serializer};
-use automerge_protocol::{ChangeHash, ObjType, Key};
+use automerge_protocol::{ChangeHash, ObjType, Key, Value};
 
 use std::collections::HashMap;
