+++ automerge-backend/src/protocol.rs
@@ -14,120 +14,8 @@ use crate::error::AutomergeError;
 use crate::helper;
 use crate::op_handle::OpHandle;
 use crate::ordered_set::OrderedSet;
-use automerge_protocol::{ActorID, ChangeHash, ObjType, OpID, ObjectID, Key};
+use automerge_protocol::{ActorID, ChangeHash, ObjType, OpID, ObjectID, Key, DataType, Value};
 
-#[derive(Serialize, PartialEq, Debug, Clone)]
-#[serde(untagged)]
-pub enum Value {
-    Str(String),
-    Int(i64),
-    Uint(u64),
-    F64(f64),
-    F32(f32),
-    Counter(i64),
-    Timestamp(i64),
-    Boolean(bool),
-    Null,
-}
-
-impl Value {
-    pub fn from(val: Option<Value>, datatype: Option<DataType>) -> Option<Value> {
-        match datatype {
-            Some(DataType::Counter) => Some(Value::Counter(val?.to_i64()?)),
-            Some(DataType::Timestamp) => Some(Value::Timestamp(val?.to_i64()?)),
-            _ => val,
-        }
-    }
-
-    pub fn adjust(self, datatype: DataType) -> Value {
-        match datatype {
-            DataType::Counter => {
-                if let Some(n) = self.to_i64() {
-                    Value::Counter(n)
-                } else {
-                    self
-                }
-            }
-            DataType::Timestamp => {
-                if let Some(n) = self.to_i64() {
-                    Value::Timestamp(n)
-                } else {
-                    self
-                }
-            }
-            _ => self,
-        }
-    }
-
-    pub fn to_i64(&self) -> Option<i64> {
-        match self {
-            Value::Int(n) => Some(*n),
-            Value::Uint(n) => Some(*n as i64),
-            Value::F32(n) => Some(*n as i64),
-            Value::F64(n) => Some(*n as i64),
-            Value::Counter(n) => Some(*n),
-            Value::Timestamp(n) => Some(*n),
-            _ => None,
-        }
-    }
-}
-
-impl From<&str> for Value {
-    fn from(s: &str) -> Self {
-        Value::Str(s.into())
-    }
-}
-
-impl From<i64> for Value {
-    fn from(n: i64) -> Self {
-        Value::Int(n)
-    }
-}
-
-impl From<u64> for Value {
-    fn from(n: u64) -> Self {
-        Value::Uint(n)
-    }
-}
-
-/*
-impl PartialOrd for ElementID {
-    fn partial_cmp(&self, other: &ElementID) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for ElementID {
-    fn cmp(&self, other: &ElementID) -> Ordering {
-        match (self, other) {
-            (ElementID::Head, ElementID::Head) => Ordering::Equal,
-            (ElementID::Head, _) => Ordering::Less,
-            (_, ElementID::Head) => Ordering::Greater,
-            (ElementID::ID(self_opid), ElementID::ID(other_opid)) => self_opid.cmp(other_opid),
-        }
-    }
-}
-*/
-
-#[derive(Deserialize, Serialize, PartialEq, Debug, Clone, Copy)]
-pub enum DataType {
-    #[serde(rename = "counter")]
-    Counter,
-    #[serde(rename = "timestamp")]
-    Timestamp,
-    #[serde(rename = "undefined")]
-    Undefined,
-}
-
-impl DataType {
-    #[allow(clippy::trivially_copy_pass_by_ref)]
-    pub fn is_undefined(d: &DataType) -> bool {
-        match d {
-            DataType::Undefined => true,
-            _ => false,
-        }
-    }
-}
 
 #[derive(Serialize, Debug, PartialEq, Clone)]
 pub enum RequestKey {
