+++ automerge-protocol/src/serde_impls/mod.rs
@@ -3,3 +3,4 @@ mod opid;
 mod object_id;
 mod element_id;
 mod key;
+mod value;
