+++ automerge-protocol/src/utility_impls/mod.rs
@@ -2,3 +2,4 @@ mod opid;
 mod object_id;
 mod key;
 mod element_id;
+mod value;
