+++ automerge-protocol/src/utility_impls/value.rs
@@ -1,0 +1,20 @@
+use crate::Value;
+
+impl From<&str> for Value {
+    fn from(s: &str) -> Self {
+        Value::Str(s.into())
+    }
+}
+
+impl From<i64> for Value {
+    fn from(n: i64) -> Self {
+        Value::Int(n)
+    }
+}
+
+impl From<u64> for Value {
+    fn from(n: u64) -> Self {
+        Value::Uint(n)
+    }
+}
+
